+++ src/printf.h
@@ -48,7 +48,7 @@ extern void err_printf(const char *fmt, ...);
 extern void fail_printf(const char *fmt, ...);
 extern void sysf_printf(const char *fmt, ...);
 
-extern void progress_bar(unsigned ctrl, unsigned x, unsigned n, unsigned w);
+extern void progress_bar(unsigned ctrl, unsigned long x, unsigned long n, unsigned w);
 
 #ifdef __cplusplus
 }
